Update README.md
This Expense Tracker project is designed to help users track their expenses effectively. It's built using ASP.NET Core for the backend and SQL Server for the database. The application allows users to register, log in, add, edit, and delete their expenses. It provides a simple and intuitive interface for managing expenses efficiently.


Features


1)User authentication system.
2)User-friendly interface for adding, editing, and deleting expenses.
3)Dashboard to view summary of expenses.